Убрал dto от цб в отдельный пакет
