+++ src/main/java/com/example/currencies/entity/cbr/Item.java
@@ -1,4 +1,4 @@
-package com.example.currencies.entity;
+package com.example.currencies.entity.cbr;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
