+++ src/main/java/com/example/currencies/entity/cbr/ValCurs.java
@@ -1,5 +1,6 @@
-package com.example.currencies.entity;
+package com.example.currencies.entity.cbr;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlElementWrapper;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlRootElement;
@@ -7,18 +8,23 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.time.LocalDate;
 import java.util.List;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-@JacksonXmlRootElement(localName = "Valuta")
-public class Valuta {
+@JacksonXmlRootElement(localName = "ValCurs")
+public class ValCurs {
 
-    @JacksonXmlProperty(localName = "name", isAttribute = true)
+    @JsonFormat(pattern = "dd.MM.yyy")
+    @JacksonXmlProperty(localName = "Date", isAttribute = true)
+    private LocalDate date;
+
+    @JacksonXmlProperty(isAttribute = true, localName = "name")
     private String name;
 
     @JacksonXmlElementWrapper(useWrapping = false)
-    @JacksonXmlProperty(localName = "Item")
-    private List<Item> items;
+    @JacksonXmlProperty(localName = "Valute")
+    private List<Valute> valutes;
 }
