+++ src/main/java/com/example/currencies/entity/cbr/Valuta.java
@@ -1,6 +1,5 @@
-package com.example.currencies.entity;
+package com.example.currencies.entity.cbr;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlElementWrapper;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlRootElement;
@@ -8,23 +7,18 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDate;
 import java.util.List;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-@JacksonXmlRootElement(localName = "ValCurs")
-public class ValCurs {
+@JacksonXmlRootElement(localName = "Valuta")
+public class Valuta {
 
-    @JsonFormat(pattern = "dd.MM.yyy")
-    @JacksonXmlProperty(localName = "Date", isAttribute = true)
-    private LocalDate date;
-
-    @JacksonXmlProperty(isAttribute = true, localName = "name")
+    @JacksonXmlProperty(localName = "name", isAttribute = true)
     private String name;
 
     @JacksonXmlElementWrapper(useWrapping = false)
-    @JacksonXmlProperty(localName = "Valute")
-    private List<Valute> valutes;
+    @JacksonXmlProperty(localName = "Item")
+    private List<Item> items;
 }
